# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Aquaventure Swim School API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /testimonials:
    get:
      operationId: listTestimonials
      tags: [testimonials]
      summary: List approved testimonials
      responses:
        "200":
          description: Approved testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
    post:
      operationId: submitTestimonial
      tags: [testimonials]
      summary: Submit a testimonial
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitTestimonialBody"
      responses:
        "201":
          description: Testimonial submitted (pending approval)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/testimonials:
    get:
      operationId: listAllTestimonials
      tags: [admin]
      summary: List all testimonials (including unapproved)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All testimonials
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Testimonial"
  /admin/testimonials/{id}/approve:
    patch:
      operationId: approveTestimonial
      tags: [admin]
      summary: Approve or unapprove a testimonial
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveTestimonialBody"
      responses:
        "200":
          description: Updated testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Testimonial"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/testimonials/{id}:
    delete:
      operationId: deleteTestimonial
      tags: [admin]
      summary: Delete a testimonial
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Testimonial:
      type: object
      properties:
        id:
          type: integer
        parentName:
          type: string
        content:
          type: string
        rating:
          type: integer
        approved:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - parentName
        - content
        - rating
        - approved
        - createdAt
    SubmitTestimonialBody:
      type: object
      properties:
        parentName:
          type: string
        content:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
      required:
        - parentName
        - content
        - rating
    ApproveTestimonialBody:
      type: object
      properties:
        approved:
          type: boolean
      required:
        - approved
...